name format fixes and renames
