+++ src/graphql/schema/create-account.graphql
@@ -1,0 +1,7 @@
+type Create {
+	newKey: String!
+}
+
+extend type Mutation {
+	createNewAccount(privateKey: String!): Create!
+}
